+++ dotnet/run_crossplatform_tests.ps1
@@ -1,5 +1,4 @@
 #!/usr/bin/env pwsh
 
-cd dotnet
-./download_cores.ps1
+./update_test.ps1
 dotnet test Tests/Eyes.Selenium.UnitTests/Eyes.Selenium.UnitTests.csproj