+++ dotnet/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Eyes.Images vNext]
+### Updated
+- Match `DeviceName` enum to devices supported by Server Core.
+
 ## [Eyes.Image.Core 3.6.1.1] - 2023-07-14
 ### Updated
 - Server Core to 3.6.1